+++ packages/api/src/tests/score.spec.ts
@@ -19,10 +19,12 @@ import { sum } from 'lodash';
 import { getPeriodDateRangeQuery } from '@period/utils';
 import faker from 'faker';
 import { add } from 'date-fns';
+import { PeriodSettingsModel } from '@periodsettings/entities';
 
 describe('calculateQuantificationScore', () => {
   beforeEach(async () => {
     await PeriodModel.deleteMany({});
+    await PeriodSettingsModel.deleteMany({});
   });
 
   it('dismissed praise scores 0', async () => {
@@ -74,6 +76,54 @@ describe('calculateQuantificationScore', () => {
     expect(score).equals(
       quantificationOriginal.score * duplicatePraisePercentage
     );
+  });
+
+  it('duplicate praise scores rounded to 2 decimal places', async () => {
+    const startDate = new Date();
+
+    const praiseOriginal = await seedPraise({
+      createdAt: startDate.setDate(startDate.getDate() + 5),
+    });
+    const praise = await seedPraise({
+      createdAt: startDate.setDate(startDate.getDate() + 5),
+    });
+    const period = await seedPeriod({
+      endDate: add(startDate, { days: 10 }),
+    });
+
+    await PeriodSettingsModel.updateOne(
+      {
+        period: period._id,
+        key: 'PRAISE_QUANTIFY_DUPLICATE_PRAISE_PERCENTAGE',
+      },
+      { $set: { value: 0.111 } }
+    );
+    const duplicatePraisePercentage = (await settingValue(
+      'PRAISE_QUANTIFY_DUPLICATE_PRAISE_PERCENTAGE',
+      period._id
+    )) as number;
+
+    const quantifier = await seedUser();
+    const quantificationOriginal = await seedQuantification(
+      praiseOriginal,
+      quantifier,
+      {
+        dismissed: false,
+        score: 144,
+      }
+    );
+    const quantification = await seedQuantification(praise, quantifier, {
+      dismissed: false,
+      duplicatePraise: praiseOriginal._id,
+    });
+
+    const score = await calculateQuantificationScore(quantification);
+
+    expect(score)
+      .equals(
+        +(quantificationOriginal.score * duplicatePraisePercentage).toFixed(2)
+      )
+      .equals(15.98);
   });
 
   it('duplicate praise scores 0 if original was dismissed', async () => {
@@ -130,9 +180,10 @@ describe('calculateQuantificationScore', () => {
 describe('calculateQuantificationsCompositeScore', () => {
   beforeEach(async () => {
     await PeriodModel.deleteMany({});
+    await PeriodSettingsModel.deleteMany({});
   });
 
-  it('composite score is floor of average of included scores', async () => {
+  it('composite score is average of included scores, rounded to 2 decimal places', async () => {
     const startDate = new Date();
 
     const praise = await seedPraise({
@@ -170,7 +221,9 @@ describe('calculateQuantificationsCompositeScore', () => {
       praise.quantifications
     );
 
-    expect(score).equals(Math.floor((10 + 30 + 50 + 70) / 4));
+    expect(score)
+      .equals(+((10 + 30 + 50 + 70) / 4).toFixed(2))
+      .equals(40.0);
   });
 
   it('incomplete quantifications are not included in composite score', async () => {
@@ -211,7 +264,9 @@ describe('calculateQuantificationsCompositeScore', () => {
       praise.quantifications
     );
 
-    expect(score).equals(Math.floor((10 + 30 + 50 + 0) / 3));
+    expect(score)
+      .equals(+((10 + 30 + 50 + 0) / 3).toFixed(2))
+      .equals(30.0);
   });
 
   it('dismissed quantifications are included in composite score', async () => {
@@ -251,7 +306,9 @@ describe('calculateQuantificationsCompositeScore', () => {
       praise.quantifications
     );
 
-    expect(score).equals(Math.floor((10 + 30 + 50 + 0) / 4));
+    expect(score)
+      .equals(+((10 + 30 + 50 + 0) / 4).toFixed(2))
+      .equals(22.5);
   });
 
   it('duplicatePraise quantifications are included in composite score', async () => {
@@ -299,7 +356,9 @@ describe('calculateQuantificationsCompositeScore', () => {
       praise.quantifications
     );
 
-    expect(score).equals(Math.floor((10 + 30 + 50 + 10) / 4));
+    expect(score)
+      .equals(+((10 + 30 + 50 + 10) / 4).toFixed(2))
+      .equals(25);
   });
 
   it('manually scored quantifications are included in composite score', async () => {
@@ -340,7 +399,66 @@ describe('calculateQuantificationsCompositeScore', () => {
       praise.quantifications
     );
 
-    expect(score).equals(Math.floor((10 + 30 + 50 + 70) / 4));
+    expect(score)
+      .equals(+((10 + 30 + 50 + 70) / 4).toFixed(2))
+      .equals(40.0);
+  });
+
+  it('composite score is rounded to 2 decimal places', async () => {
+    const startDate = new Date();
+
+    const praiseOriginal = await seedPraise({
+      createdAt: startDate.setDate(startDate.getDate() + 5),
+    });
+    const praise = await seedPraise({
+      createdAt: startDate.setDate(startDate.getDate() + 5),
+    });
+    const period = await seedPeriod({
+      endDate: add(startDate, { days: 10 }),
+    });
+    await PeriodSettingsModel.updateOne(
+      {
+        period: period._id,
+        key: 'PRAISE_QUANTIFY_DUPLICATE_PRAISE_PERCENTAGE',
+      },
+      { $set: { value: 0.111 } }
+    );
+
+    const quantifier = await seedUser();
+    await seedQuantification(praise, quantifier, {
+      dismissed: false,
+      score: 10,
+    });
+
+    const quantifier2 = await seedUser();
+    await seedQuantification(praise, quantifier2, {
+      dismissed: false,
+      score: 30,
+    });
+
+    const quantifier3 = await seedUser();
+    await seedQuantification(praise, quantifier3, {
+      dismissed: false,
+      score: 50,
+    });
+
+    const quantifier4 = await seedUser();
+    await seedQuantification(praiseOriginal, quantifier4, {
+      dismissed: false,
+      score: 144,
+    });
+    await seedQuantification(praise, quantifier4, {
+      dismissed: false,
+      duplicatePraise: praiseOriginal._id,
+    });
+
+    const score = await calculateQuantificationsCompositeScore(
+      praise.quantifications
+    );
+
+    expect(score)
+      .equals(+((10 + 30 + 50 + 15.98) / 4).toFixed(2))
+      .equals(26.5);
   });
 });
 
@@ -348,6 +466,7 @@ describe('calculateReceiverCompositeScore', () => {
   beforeEach(async () => {
     await PeriodModel.deleteMany({});
     await PraiseModel.deleteMany({});
+    await PeriodSettingsModel.deleteMany({});
   });
 
   it('receiver composite score is sum of all quantification composite scores', async () => {
@@ -488,6 +607,6 @@ describe('calculateReceiverCompositeScore', () => {
       )
     );
 
-    expect(score).equals(sum(compositeScores));
+    expect(score).equals(+sum(compositeScores).toFixed(2));
   });
 });