+++ packages/discord-bot/src/handlers/activate.ts
@@ -4,8 +4,24 @@ import { EventLogTypeKey } from 'api/src/eventlog/types';
 import { logEvent } from 'api/src/eventlog/utils';
 import randomstring from 'randomstring';
 import { CommandHandler } from 'src/interfaces/CommandHandler';
+import { alreadyActivatedError } from '../utils/praiseEmbeds';
+
 export const activationHandler: CommandHandler = async (interaction) => {
   const { user } = interaction;
+
+  let userAccount = await UserAccountModel.findOne({
+    accountId: user.id,
+    platform: 'DISCORD',
+  });
+
+  if (userAccount?.user) {
+    await interaction.reply({
+      content: await alreadyActivatedError(),
+      ephemeral: true,
+    });
+    return;
+  }
+
   const ua = {
     accountId: user.id,
     name: user.username + '#' + user.discriminator,
@@ -13,7 +29,8 @@ export const activationHandler: CommandHandler = async (interaction) => {
     platform: 'DISCORD',
     activateToken: randomstring.generate(),
   } as UserAccount;
-  const userAccount = await UserAccountModel.findOneAndUpdate(
+
+  userAccount = await UserAccountModel.findOneAndUpdate(
     { accountId: user.id },
     ua,
     { upsert: true, new: true }